# MTD Validation: SIM clusters