--- a/Validation/MtdValidation/plugins/BtlLocalRecoValidation.cc
+++ b/Validation/MtdValidation/plugins/BtlLocalRecoValidation.cc
@@ -131,6 +131,8 @@ private:
   MonitorElement* meCluEnergyRes_;
   MonitorElement* meCluTPullvsE_;
   MonitorElement* meCluTPullvsEta_;
+  MonitorElement* meCluRhoRes_;
+  MonitorElement* meCluPhiRes_;
   MonitorElement* meCluXRes_;
   MonitorElement* meCluYRes_;
   MonitorElement* meCluZRes_;
@@ -311,9 +313,7 @@ void BtlLocalRecoValidation::analyze(const edm::Event& iEvent, const edm::EventS
       const RectangularMTDTopology& topo = static_cast<const RectangularMTDTopology&>(topoproxy.specificTopology());
 
       // --- Cluster position in the module reference frame
-      Local3DPoint local_point((cluster.x() - 0.5 * topo.nrows()) * topo.pitch().first,
-                               (cluster.y() - 0.5 * topo.ncolumns()) * topo.pitch().second,
-                               0.);
+      Local3DPoint local_point(topo.localX(cluster.x()), topo.localY(cluster.y()), 0.);
       const auto& global_point = genericDet->toGlobal(local_point);
 
       meCluEnergy_->Fill(cluster.energy());
@@ -351,11 +351,11 @@ void BtlLocalRecoValidation::analyze(const edm::Event& iEvent, const edm::EventS
             continue;
 
           // SIM hit's position in the module reference frame
-          Local3DPoint local_point_sim(convertMmToCm(m_btlSimHits[recHit.id().rawId()].x_local) +
-                                           (hit_row - 0.5 * topo.nrows() + 0.5) * topo.pitch().first,
-                                       convertMmToCm(m_btlSimHits[recHit.id().rawId()].y_local) +
-                                           (hit_col - 0.5 * topo.ncolumns() + 0.5) * topo.pitch().second,
+          Local3DPoint local_point_sim(convertMmToCm(m_btlSimHits[recHit.id().rawId()].x_local),
+                                       convertMmToCm(m_btlSimHits[recHit.id().rawId()].y_local),
                                        convertMmToCm(m_btlSimHits[recHit.id().rawId()].z_local));
+          local_point_sim =
+              topo.pixelToModuleLocalPoint(local_point_sim, hitId.row(topo.nrows()), hitId.column(topo.nrows()));
 
           // Calculate the SIM cluster's position in the module reference frame
           cluLocXSIM += local_point_sim.x() * m_btlSimHits[recHit.id().rawId()].energy;
@@ -379,15 +379,24 @@ void BtlLocalRecoValidation::analyze(const edm::Event& iEvent, const edm::EventS
 
         float time_res = cluster.time() - cluTimeSIM;
         float energy_res = cluster.energy() - cluEneSIM;
-        float x_res = global_point.x() - cluGlobalPosSIM.x();
-        float y_res = global_point.y() - cluGlobalPosSIM.y();
-        float z_res = global_point.z() - cluGlobalPosSIM.z();
-
         meCluTimeRes_->Fill(time_res);
         meCluEnergyRes_->Fill(energy_res);
-        meCluXRes_->Fill(x_res);
-        meCluYRes_->Fill(y_res);
-        meCluZRes_->Fill(z_res);
+
+        float rho_res = global_point.perp() - cluGlobalPosSIM.perp();
+        float phi_res = global_point.phi() - cluGlobalPosSIM.phi();
+
+        meCluRhoRes_->Fill(rho_res);
+        meCluPhiRes_->Fill(phi_res);
+
+        if (LocalPosDebug_) {
+          float x_res = global_point.x() - cluGlobalPosSIM.x();
+          float y_res = global_point.y() - cluGlobalPosSIM.y();
+          float z_res = global_point.z() - cluGlobalPosSIM.z();
+
+          meCluXRes_->Fill(x_res);
+          meCluYRes_->Fill(y_res);
+          meCluZRes_->Fill(z_res);
+        }
 
         meCluTPullvsEta_->Fill(std::abs(cluGlobalPosSIM.eta()), time_res / cluster.timeError());
         meCluTPullvsE_->Fill(cluEneSIM, time_res / cluster.timeError());
@@ -575,8 +584,9 @@ void BtlLocalRecoValidation::bookHistograms(DQMStore::IBooker& ibook,
   meCluZvsPhi_ = ibook.book2D(
       "BtlOccupancy", "BTL cluster Z vs #phi;Z_{RECO} [cm]; #phi_{RECO} [rad]", 144, -260., 260., 50, -3.2, 3.2);
 
-  meCluTimeRes_ = ibook.book1D("BtlCluTimeRes", "BTL cluster time resolution;T_{RECO}-T_{SIM}", 100, -0.5, 0.5);
-  meCluEnergyRes_ = ibook.book1D("BtlCluEnergyRes", "BTL cluster energy resolution;E_{RECO}-E_{SIM}", 100, -0.5, 0.5);
+  meCluTimeRes_ = ibook.book1D("BtlCluTimeRes", "BTL cluster time resolution;T_{RECO}-T_{SIM} [ns]", 100, -0.5, 0.5);
+  meCluEnergyRes_ =
+      ibook.book1D("BtlCluEnergyRes", "BTL cluster energy resolution;E_{RECO}-E_{SIM} [MeV]", 100, -0.5, 0.5);
   meCluTPullvsE_ = ibook.bookProfile("BtlCluTPullvsE",
                                      "BTL cluster time pull vs E;E_{SIM} [MeV];(T_{RECO}-T_{SIM})/#sigma_{T_{RECO}}",
                                      20,
@@ -594,9 +604,15 @@ void BtlLocalRecoValidation::bookHistograms(DQMStore::IBooker& ibook,
                         -5.,
                         5.,
                         "S");
-  meCluXRes_ = ibook.book1D("BtlCluXRes", "BTL cluster X resolution;X_{RECO}-X_{SIM} [cm]", 100, -3.1, 3.1);
-  meCluYRes_ = ibook.book1D("BtlCluYRes", "BTL cluster Y resolution;Y_{RECO}-Y_{SIM} [cm]", 100, -3.1, 3.1);
-  meCluZRes_ = ibook.book1D("BtlCluZRes", "BTL cluster Z resolution;Z_{RECO}-Z_{SIM} [cm]", 100, -0.2, 0.2);
+  meCluRhoRes_ =
+      ibook.book1D("BtlCluRhoRes", "BTL cluster #rho resolution;#rho_{RECO}-#rho_{SIM} [cm]", 100, -0.5, 0.5);
+  meCluPhiRes_ =
+      ibook.book1D("BtlCluPhiRes", "BTL cluster #phi resolution;#phi_{RECO}-#phi_{SIM} [rad]", 100, -0.03, 0.03);
+  if (LocalPosDebug_) {
+    meCluXRes_ = ibook.book1D("BtlCluXRes", "BTL cluster X resolution;X_{RECO}-X_{SIM} [cm]", 100, -3.1, 3.1);
+    meCluYRes_ = ibook.book1D("BtlCluYRes", "BTL cluster Y resolution;Y_{RECO}-Y_{SIM} [cm]", 100, -3.1, 3.1);
+    meCluZRes_ = ibook.book1D("BtlCluZRes", "BTL cluster Z resolution;Z_{RECO}-Z_{SIM} [cm]", 100, -0.2, 0.2);
+  }
   meCluYXLocal_ = ibook.book2D("BtlCluYXLocal",
                                "BTL cluster local Y vs X;X^{local}_{RECO} [cm];Y^{local}_{RECO} [cm]",
                                200,