@fabiocos 2021-06-22
@casarsa do you have the output plots to attach as documentation? In particular those showing the behaviour studied in the uncertainties tuning?

Do you always manage to find the simhits associated to the cluster? Do you have a check for this?