@casarsa 2021-06-23
> Do you always manage to find the simhits associated to the cluster? Do you have a check for this?

I've added a check. From a quick check: out of ~380 BTL reco hits and ~370 reco ETL I always have a match to a sim hit.