@cmsbuild 2021-06-23
+1

**Summary**: https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-133f91/16189/summary.html
**COMMIT**: 482a96dbf0c114cb2726e52cbb8c771c067a6b18
**CMSSW**: CMSSW_12_0_X_2021-06-22-2300/slc7_amd64_gcc900
**User test area**: For local testing, you can use `/cvmfs/cms-ci.cern.ch/week0/cms-sw/cmssw/34207/16189/install.sh` to create a dev area with all the needed externals and cmssw changes.


## Comparison Summary



**Summary**:
 - No significant changes to the logs found
 - Reco comparison results: 4 differences found in the comparisons
 - DQMHistoTests: Total files compared: 38
 - DQMHistoTests: Total histograms compared: 2785631
 - DQMHistoTests: Total failures: 23
 - DQMHistoTests: Total nulls: 4
 - DQMHistoTests: Total successes: 2785582
 - DQMHistoTests: Total skipped: 22
 - DQMHistoTests: Total Missing objects: 0
 - DQMHistoSizes: Histogram memory added: 30.256 KiB( 37 files compared)
 - DQMHistoSizes: changed ( 23234.0,... ): 5.379 KiB        MTD/ETL
 - DQMHistoSizes: changed ( 23234.0,... ): 2.186 KiB        MTD/BTL
 - Checked 160 log files, 37 edm output root files, 38 DQM output files
 - [TriggerResults](https://cmssdt.cern.ch/SDT/jenkins-artifacts/baseLineComparisons/CMSSW_12_0_X_2021-06-22-2300+133f91/43625/triggerResults): no differences found