@casarsa 2021-06-24
> @casarsa do you understand the pattern of the differences? Why no resolution plot and ETL plot? Didi you see them in the local test output?

Yes, I think I do:
BtlLongposPullvsEta: I changed the SIM hit position in the crystal from mm to cm and this affects the eta.

BtlCluPhi and BtlCluEta: I fixed the calculation of the cluster local position in the module. RectangularMTDTopology::pixelToModuleLocalPoint was used to transform coordinates that already were in the module frame.

BtlTPullvsEta: I've changed the number of bins for uniformity with other similar histograms.