@fabiocos 2021-06-24
@casarsa thanks, the new histograms are indeed available in the output